Refactor type imports and clean referral data
Updated type imports in IndividualInfoSection and ReferralDetailsSection to use the centralized types module. Added logic in APIService to remove empty string values from referral data before sending to the backend, ensuring optional fields are handled as undefined. Also set the date input type for the date of birth field.
